Python/TicketmastersEventFinderAPI/requirements.txt
fastapi
requests
pandas
python-dotenv
uvicorn
